remove the test file
